+++ src/main/java/com/pisip/jbpharma/infraestructura/repositorio/iEnsayoLaboratoriojpaRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pisip.jbpharma.infraestructura.persistencia.jpa.EnsayoLaboratorioEntity;
 
-public interface iEnsayoLaboratoriojpaRepositorio extends JpaRepository<EnsayoLaboratorioEntity, Integer>{
+public interface iEnsayoLaboratoriojpaRepositorio extends JpaRepository<EnsayoLaboratorioEntity, Long>{
 
 }
