+++ src/main/java/com/pisip/jbpharma/infraestructura/repositorio/iEnsayoVariablejpaRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pisip.jbpharma.infraestructura.persistencia.jpa.EnsayoVariableEntity;
 
-public interface iEnsayoVariablejpaRepositorio extends JpaRepository<EnsayoVariableEntity,Integer>{
+public interface iEnsayoVariablejpaRepositorio extends JpaRepository<EnsayoVariableEntity, Long>{
 
 }
